Update Prometheus extension to the new extension model
