+++ prometheus-extension/native/src/main/resources/META-INF/services/io.ballerina.runtime.observability.metrics.spi.MetricReporterFactory
@@ -1,0 +1,1 @@
+org.ballerinalang.observe.metrics.prometheus.PrometheusMetricsReporterFactory
