--- prometheus-extension/native/src/main/resources/META-INF/services/org.ballerinalang.jvm.observability.metrics.spi.MetricReporter
@@ -1,1 +1,0 @@
-org.ballerinalang.observe.metrics.prometheus.PrometheusReporter
